pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "my_project"
dynamic = ["version"]
description = "TBD"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "loguru<1.0.0",
  "cowsay"
]

[project.optional-dependencies]
dev = [
  "databricks-connect==15.4.*",
  "ipython==8.26.*",
  "ruff==0.8.*",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  # pycodestyle
  "E",
  # Pyflakes
  "F",
  # pyupgrade
  "UP",
  # flake8-bugbear
  "B",
  # flake8-simplify
  "SIM",
  # isort
  "I",
]

exclude = ["notebooks/*"]